+++ packages/react/src/runtimes/core/ThreadRuntime.tsx
@@ -1,3 +1,4 @@
+import { Attachment } from "../../context/stores/Attachment";
 import { RuntimeCapabilities } from "../../context/stores/Thread";
 import { ThreadActionsState } from "../../context/stores/ThreadActions";
 import { ThreadMessage } from "../../types";
@@ -7,6 +8,7 @@ export type ThreadRuntime = ThreadActionsState &
   Readonly<{
     composer: ThreadRuntime.Composer;
     capabilities: Readonly<RuntimeCapabilities>;
+    threadId: string;
     isDisabled: boolean;
     messages: readonly ThreadMessage[];
     subscribe: (callback: () => void) => Unsubscribe;
@@ -14,7 +16,15 @@ export type ThreadRuntime = ThreadActionsState &
 
 export declare namespace ThreadRuntime {
   export type Composer = Readonly<{
+    attachments: Attachment[];
+    addAttachment: (file: File) => Promise<void>;
+    removeAttachment: (attachmentId: string) => Promise<void>;
+
     text: string;
     setText: (value: string) => void;
+
+    reset: () => void;
+
+    send: () => void;
   }>;
 }