+++ packages/react/src/runtimes/index.ts
@@ -3,3 +3,5 @@ export * from "./local";
 export * from "./edge";
 export * from "./external-store";
 export * from "./dangerous-in-browser";
+export * from "./speech";
+export * from "./attachment";