+++ packages/react/src/ui/assistant-message.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ComponentPropsWithoutRef, forwardRef, type FC } from "react";
+import { ComponentPropsWithoutRef, forwardRef, useMemo, type FC } from "react";
 import { MessagePrimitive, MessagePrimitiveContentProps } from "../primitives";
 import BranchPicker from "./branch-picker";
 import { Avatar } from "./base/avatar";
@@ -44,13 +44,32 @@ const AssistantMessageContent = forwardRef<
   HTMLDivElement,
   AssistantMessageContentProps
 >(({ components: componentsProp, ...rest }, ref) => {
-  const { assistantMessage: { components = {} } = {} } = useThreadConfig();
+  const { tools, assistantMessage: { components = {} } = {} } =
+    useThreadConfig();
+
+  const toolsComponents = useMemo(
+    () => ({
+      by_name: !tools
+        ? undefined
+        : Object.fromEntries(
+            tools.map((t) => [
+              t.unstable_tool.toolName,
+              t.unstable_tool.render,
+            ]),
+          ),
+      Fallback: components.ToolFallback,
+    }),
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [...(tools ?? []), components.ToolFallback],
+  );
+
   return (
     <AssistantMessageContentWrapper {...rest} ref={ref}>
       <MessagePrimitive.Content
         components={{
           ...componentsProp,
           Text: componentsProp?.Text ?? components.Text ?? ContentPart.Text,
+          tools: toolsComponents,
         }}
       />
     </AssistantMessageContentWrapper>