+++ packages/shadcn-registry/registry/assistant-ui/thread.tsx
@@ -11,28 +11,28 @@ import { SendHorizontalIcon } from "lucide-react";
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { TooltipIconButton } from "@/components/ui/assistant-ui/tooltip-icon-button";
 
-export const Thread: FC = () => {
+export const MyThread: FC = () => {
   return (
     <ThreadPrimitive.Root className="bg-background h-full">
       <ThreadPrimitive.Viewport className="flex h-full flex-col items-center overflow-y-scroll scroll-smooth bg-inherit px-4 pt-8">
-        <ThreadWelcome />
+        <MyThreadWelcome />
 
         <ThreadPrimitive.Messages
           components={{
-            UserMessage,
-            AssistantMessage,
+            UserMessage: MyUserMessage,
+            AssistantMessage: MyAssistantMessage,
           }}
         />
 
         <div className="sticky bottom-0 mt-4 flex w-full max-w-2xl flex-grow flex-col items-center justify-end rounded-t-lg bg-inherit pb-4">
-          <Composer />
+          <MyComposer />
         </div>
       </ThreadPrimitive.Viewport>
     </ThreadPrimitive.Root>
   );
 };
 
-const ThreadWelcome: FC = () => {
+const MyThreadWelcome: FC = () => {
   return (
     <ThreadPrimitive.Empty>
       <div className="flex flex-grow basis-full flex-col items-center justify-center">
@@ -45,20 +45,20 @@ const ThreadWelcome: FC = () => {
   );
 };
 
-const Composer: FC = () => {
+const MyComposer: FC = () => {
   return (
-    <ComposerPrimitive.Root className="relative flex w-full items-end rounded-lg border transition-shadow focus-within:shadow-sm">
+    <ComposerPrimitive.Root className="focus-within:border-aui-ring/20 flex w-full flex-wrap items-end rounded-lg border px-2.5 shadow-sm transition-colors ease-in">
       <ComposerPrimitive.Input
         autoFocus
         placeholder="Write a message..."
         rows={1}
-        className="placeholder:text-muted-foreground size-full max-h-40 resize-none border-none bg-transparent p-4 pr-12 text-sm outline-none focus:ring-0"
+        className="placeholder:text-muted-foreground size-full max-h-40 resize-none border-none bg-transparent p-4 pr-12 text-sm outline-none focus:ring-0 disabled:cursor-not-allowed"
       />
       <ComposerPrimitive.Send asChild>
         <TooltipIconButton
           tooltip="Send"
           variant="default"
-          className="absolute bottom-0 right-0 m-2.5 size-8 p-2 transition-opacity"
+          className="my-2.5 size-8 p-2 transition-opacity ease-in"
         >
           <SendHorizontalIcon />
         </TooltipIconButton>
@@ -67,7 +67,7 @@ const Composer: FC = () => {
   );
 };
 
-const UserMessage: FC = () => {
+const MyUserMessage: FC = () => {
   return (
     <MessagePrimitive.Root className="grid w-full max-w-2xl auto-rows-auto grid-cols-[minmax(72px,1fr)_auto] gap-y-2 py-4">
       <div className="bg-muted text-foreground col-start-2 row-start-1 max-w-xl break-words rounded-3xl px-5 py-2.5">
@@ -77,7 +77,7 @@ const UserMessage: FC = () => {
   );
 };
 
-const AssistantMessage: FC = () => {
+const MyAssistantMessage: FC = () => {
   return (
     <MessagePrimitive.Root className="relative grid w-full max-w-2xl grid-cols-[auto_1fr] grid-rows-[auto_1fr] py-4">
       <Avatar className="col-start-1 row-span-full row-start-1 mr-4">