+++ packages/react-playground/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @assistant-ui/react-playground
 
+## 0.0.23
+
+### Patch Changes
+
+- fb8e58f: feat: add thread runtime threadId
+- Updated dependencies [fb8e58f]
+  - @assistant-ui/react@0.5.45
+
+## 0.0.22
+
+### Patch Changes
+
+- c8b98b6: feat: animate composer border color on focus
+- Updated dependencies [c8b98b6]
+- Updated dependencies [800eb9e]
+- Updated dependencies [8768c67]
+  - @assistant-ui/react@0.5.42
+
+## 0.0.21
+
+### Patch Changes
+
+- f526279: feat: SpeechSyntehsis
+- e8aa697: refactor: remove unsupported external runtime onCopy callback
+- Updated dependencies [f526279]
+- Updated dependencies [e8aa697]
+  - @assistant-ui/react@0.5.41
+
 ## 0.0.20
 
 ### Patch Changes